requirements.txt
# Core FastAPI and Web Framework
fastapi==0.117.1
uvicorn==0.36.0
python-multipart==0.0.20
pydantic[dotenv]==2.11.9
watchfiles==1.1.0

# Database
sqlmodel==0.0.25
psycopg2-binary==2.9.10

# Basic utilities
python-dotenv==1.1.1
aiofiles==24.1.0
requests==2.32.5
setuptools==80.9.0

# Image processing and computer vision
opencv-python==4.10.0.84
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.3.0
numpy==2.3.3
scikit-image==0.25.2
imageio==2.37.0

# OCR libraries
pytesseract==0.3.13
paddlepaddle==3.0.0b2
paddleocr==2.9.1

# ML and Data Processing
scikit-learn==1.7.2
scipy==1.16.2
albumentations==1.4.10
imgaug==0.4.0

# Document processing
python-docx==1.2.0
lxml==6.0.1
beautifulsoup4==4.13.5
pdf2image==1.17.0

# Text processing and NLP utilities
RapidFuzz==3.14.1
pyclipper==1.3.0.post6

# Visualization
matplotlib==3.10.6

# Additional ML utilities
joblib==1.5.2
tqdm==4.67.1
PyYAML==6.0.2
protobuf==6.32.1

# Development and testing
httpx==0.28.1
pytest==8.3.3
pytest-asyncio==0.24.0

# System utilities
fire==0.7.1
termcolor==3.1.0
Cython==3.1.4